project_files/task3/requirements.txt
click==8.0.1
dataclasses==0.8
Flask==2.0.1
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
psutil==5.8.0
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0